Clean up.
Deleted unused MapContainer. Updated javadoc.

+++ src/de/fhpotsdam/unfolding/marker/SimplePointMarker.java
@@ -2,6 +2,7 @@ package de.fhpotsdam.unfolding.marker;
 
 import java.util.HashMap;
 
+import processing.core.PApplet;
 import processing.core.PGraphics;
 import processing.core.PVector;
 import de.fhpotsdam.unfolding.geo.Location;
@@ -57,7 +58,7 @@ public class SimplePointMarker extends AbstractMarker {
 			pg.fill(color);
 			pg.stroke(strokeColor);
 		}
-		pg.ellipse(x, y, radius, radius); // TODO use radius in km and convert to px
+		pg.ellipse((int)x, (int)y, radius, radius); // TODO use radius in km and convert to px
 		pg.popStyle();
 	}
